updated job.Bid.MinRate from float32 to float64

+++ pkg/job/Bid.go
@@ -3,5 +3,5 @@ package job
 // Bid represents a minimum price or rate a miner is willing to accept to execute a job
 type Bid struct {
 	JobID   string
-	MinRate float32
+	MinRate float64
 }
